backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
psycopg2-binary
pandas
numpy
scikit-learn
xgboost
joblib
python-dotenv
marshmallow
flask-marshmallow
